get user specific tour details and update tour data

+++ travxy/resources/tour.py
@@ -2,6 +2,7 @@ from flask_restful import Resource, request
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from travxy.models.tour import TourModel
 from travxy.models.tourist import TouristInfoModel
+from travxy.models.category import CategoryModel
 from travxy.helpers.pagination import get_paginated_list
 from sqlalchemy.orm import joinedload
 
@@ -16,6 +17,10 @@ class Tour(Resource):
 
     @jwt_required()
     def delete(self, tour_id):
+        user_id = get_jwt_identity()
+        current_user = TouristInfoModel.find_by_user_id(user_id)
+        if current_user.role_id != 1 and current_user.role_id != 2:
+            return {'message': 'Unauthorized User'}, 401
         tour = TourModel.find_by_id(tour_id)
         if tour:
             tour.delete_from_db()
@@ -30,7 +35,7 @@ class TourList(Resource):
                                         start=request.args.get('start', 1),
                                         limit=request.args.get('limit', 20))
 
-class SearchTourDetail(Resource):
+class SearchTours(Resource):
     @jwt_required()
     def get(self, search_term):
         current_identity = get_jwt_identity()
@@ -44,10 +49,10 @@ class SearchTourDetail(Resource):
                         '%'+search_term+'%')).all()
         if not results:
             return {'message': 'Tour does not exist'}, 400
-        for result in results:
-            return result.with_category_json()
+        tours = [result.with_category_json() for result in results]
+        return tours
 
-class AdminForTour(Resource):
+class AdminForNewTours(Resource):
     @jwt_required()
     def post(self):
         user_id = get_jwt_identity()
@@ -55,23 +60,29 @@ class AdminForTour(Resource):
         if current_user.role_id != 1 and current_user.role_id != 2:
             return {'message': 'Unauthorized User'}, 401
 
-        name = request.json.get('tour_name')
+        name = request.json.get('name')
         location = request.json.get('location')
         country = request.json.get('country')
         about = request.json.get('about')
+
         category_id = request.json.get('category_id')
 
-        if not all([name, location, country, about, category_id]):
+        if not all([name, location, country, about]):
             return {'message': 'Missing fields required'}, 400
         tour = TourModel(name=name, location=location, country=country,
-                            about=about, category_id=category_id)
+                        about=about)
+        category_instance = CategoryModel.find_by_id(category_id)
+        if category_instance is None:
+            return {'message': 'Category does not exist for this tour'}, 400
+        category_instance.tour_details.append(tour)
         try:
-            tour.save_to_db()
+            category_instance.save_to_db()
         except:
             return{'message':
                     'An error occured while trying to insert the tour'}, 500
-        return tour.json(), 201
+        return tour.with_category_json(), 201
 
+class AdminForSameCategoryTours(Resource):
     @jwt_required()
     def put(self):
         user_id = get_jwt_identity()
@@ -85,20 +96,27 @@ class AdminForTour(Resource):
         about = request.json.get('about')
         category_id = request.json.get('category_id')
 
-        if not all([name, location, country, about, category_id]):
+        if not all([tour_id, name, location, country, about]):
             return {'message': 'Missing fields required'}, 400
         tour = TourModel.find_by_id(tour_id)
+        same_category_instance = CategoryModel.query.join(TourModel, 
+                                CategoryModel.tour_details).filter(
+                                CategoryModel.id==category_id).filter(
+                                TourModel.id==tour_id).first()
         if tour is None:
             return {'message': 'Tour does not exist'}
-        else:
-           tour.tour_id = tour_id
-           tour.name = name
-           tour.location = location
-           tour.country = country
-           tour.about = about
-           tour.category_id = category_id
-        tour.save_to_db()
+        if same_category_instance:
+            tour.tour_id = tour_id
+            tour.name = name
+            tour.location = location
+            tour.country = country
+            tour.about = about
+            for tour_category in tour.categories_info:
+                tour.categories_info.remove(tour_category)
+                tour.categories_info.append(same_category_instance)
+        try:
+            tour.save_to_db()
+        except:
+            return {'message':
+                     'An error occured while trying to update the tour'}, 500
         return tour.json()
-
-
-
